Add support for propel-events from GlorpenPropelBundle.
I've bumped into problem when creating custom driver - when using `propel` orm option,
`getUploadableClasses` and `getUploadableFields` are called when Container is not yet built,
and because of it some base services are unavailable - eg. `kernel`.

This PR adds support for GlorpenPropelBundle as another "orm": `propel_ge`, when used allows to skip
reading data from not-yet-built Container in `RegisterPropelModelsPass`.
Going further, event handling looks more like in Doctrine part of this bundle and thus enables a way
to write custom drivers when using Propel.